dev_requirements.txt
django==3.1.14
django-bootstrap-pagination==1.7.1
django-cors-headers==3.7.0
django-crispy-forms==1.13.0
django-debug-toolbar==3.2.1
djangorestframework==3.12.2
ipython==7.16.3
markdown==3.1.1
solrpy==0.9.9
wheel==0.38.1

# testing
mixer==7.2.2
Faker==12.0.1
model-bakery==1.10.1
playwright==1.30.0
coverage==7.1.0



# python==3.8.0
# solr==8.3.0
# swig==4.0.1
# verovio==2.3.3
# psycopg2==2.8.4
